+++ menu/__init__.py
@@ -6,7 +6,23 @@ class MainPanel:
 
 from . import menu_select_model
 from . import menu_warnings
+
+
+# Configure panel
 from . import menu_configure
+# Sub-panels
+from . import menu_configure_body
+from . import menu_configure_outfit
+from . import menu_configure_hair
+from . import menu_configure_armature
+from . import menu_configure_physics
+from . import menu_configure_tools
+from . import menu_configure_external
+from . import menu_configure_links
+from . import menu_configure_others
+from . import menu_configure_debug
+from . import menu_configure_complete
+
 from . import menu_tools_creators
 from . import menu_body
 from . import menu_morphs
@@ -25,6 +41,17 @@ def register():
     menu_select_model.register()
     menu_warnings.register()
     menu_configure.register()
+    menu_configure_body.register()
+    menu_configure_outfit.register()
+    menu_configure_hair.register()
+    menu_configure_armature.register()
+    menu_configure_physics.register()
+    menu_configure_tools.register()
+    menu_configure_external.register()
+    menu_configure_links.register()
+    menu_configure_others.register()
+    menu_configure_debug.register()
+    menu_configure_complete.register()
     menu_tools_creators.register()
     menu_body.register()
     menu_morphs.register()
@@ -52,6 +79,17 @@ def unregister():
     menu_morphs.unregister()
     menu_body.unregister()
     menu_tools_creators.unregister()
+    menu_configure_complete.unregister()
+    menu_configure_debug.unregister()
+    menu_configure_others.unregister()
+    menu_configure_links.unregister()
+    menu_configure_external.unregister()
+    menu_configure_tools.unregister()
+    menu_configure_physics.unregister()
+    menu_configure_armature.unregister()
+    menu_configure_hair.unregister()
+    menu_configure_outfit.unregister()
+    menu_configure_body.unregister()
     menu_configure.unregister()
     menu_warnings.unregister()
     menu_select_model.unregister()
